+++ core/core-database/src/main/kotlin/com/xquare/coredatabase/di/DatabaseModule.kt
@@ -1,0 +1,5 @@
+package com.xquare.coredatabase.di
+
+import org.koin.dsl.module
+
+val databaseModule = module {}
