--- core/core-database/src/main/kotlin/com/xquare/database/di/DatabaseModule.kt
@@ -1,5 +1,0 @@
-package com.xquare.database.di
-
-import org.koin.dsl.module
-
-val databaseModule = module {}
