+++ core/core-network/src/main/kotlin/com/xquare/corenetwork/di/NetworkModule.kt
@@ -1,0 +1,8 @@
+package com.xquare.corenetwork.di
+
+import com.xquare.corenetwork.interceptor.AuthorizationInterceptor
+import org.koin.dsl.module
+
+val networkModule = module {
+    single { AuthorizationInterceptor() }
+}
