+++ core/core-network/src/main/kotlin/com/xquare/corenetwork/interceptor/AuthorizationInterceptor.kt
@@ -1,0 +1,10 @@
+package com.xquare.corenetwork.interceptor
+
+import okhttp3.Interceptor
+import okhttp3.Response
+
+internal class AuthorizationInterceptor : Interceptor {
+    override fun intercept(chain: Interceptor.Chain): Response {
+        TODO("Not yet implemented")
+    }
+}
