--- core/core-network/src/main/kotlin/com/xquare/network/di/NetworkModule.kt
@@ -1,8 +1,0 @@
-package com.xquare.network.di
-
-import com.xquare.network.interceptor.AuthorizationInterceptor
-import org.koin.dsl.module
-
-val networkModule = module {
-    single { AuthorizationInterceptor() }
-}
