--- core/core-network/src/main/kotlin/com/xquare/network/interceptor/AuthorizationInterceptor.kt
@@ -1,10 +1,0 @@
-package com.xquare.network.interceptor
-
-import okhttp3.Interceptor
-import okhttp3.Response
-
-internal class AuthorizationInterceptor : Interceptor {
-    override fun intercept(chain: Interceptor.Chain): Response {
-        TODO("Not yet implemented")
-    }
-}
