--- feature/main/feature-main-all/src/main/kotlin/com/xquare/feature_main_all/ui/Dummy.kt
@@ -1,1 +1,0 @@
-package com.xquare.feature_main_all.ui
