+++ feature/main/feature-main-all/src/main/kotlin/com/xquare/featuremainall/ui/Dummy.kt
@@ -1,0 +1,1 @@
+package com.xquare.featuremainall.ui
