--- feature/main/feature-main-application/src/main/java/com/xquare/feature_main_application/ui/Dummy.kt
@@ -1,1 +1,0 @@
-package com.xquare.feature_main_application.ui
