+++ feature/main/feature-main-application/src/main/kotlin/com/xquare/featuremainapplication/data/Dummy.kt
@@ -1,0 +1,1 @@
+package com.xquare.featuremainapplication.data
