+++ feature/main/feature-main-application/src/main/kotlin/com/xquare/featuremainapplication/ui/Dummy.kt
@@ -1,0 +1,1 @@
+package com.xquare.featuremainapplication.ui
