--- feature/main/feature-main-feed/src/main/kotlin/com/xqaure/feature_main_feed/ui/Dummy.kt
@@ -1,1 +1,0 @@
-package com.xqaure.feature_main_feed.ui
