+++ feature/main/feature-main-feed/src/main/kotlin/com/xqaure/featuremainfeed/data/Dummy.kt
@@ -1,0 +1,1 @@
+package com.xqaure.featuremainfeed.data
