+++ feature/main/feature-main-home/src/main/kotlin/com/xquare/featuremainhome/data/Dummy.kt
@@ -1,0 +1,1 @@
+package com.xquare.featuremainapplication.data
