--- feature/main/feature-main-schedule/src/main/kotlin/com/xquare/feature_main_schedule/data/Dummy.kt
@@ -1,1 +1,0 @@
-package com.xquare.feature_main_application.data
