+++ feature/main/feature-main-schedule/src/main/kotlin/com/xquare/featuremainschedule/data/Dummy.kt
@@ -1,0 +1,1 @@
+package com.xquare.featuremainapplication.data
