--- feature/main/feature-main/src/main/kotlin/com/xquare/feature_main/data/Dummy.kt
@@ -1,1 +1,0 @@
-package com.xquare.feature_main_application.data
