--- feature/main/feature-main/src/main/kotlin/com/xquare/feature_main/ui/Dummy.kt
@@ -1,1 +1,0 @@
-package com.xquare.feature_main_application.ui
