+++ feature/main/feature-main/src/main/kotlin/com/xquare/featuremain/data/Dummy.kt
@@ -1,0 +1,1 @@
+package com.xquare.featuremainapplication.data
