+++ feature/main/feature-main/src/main/kotlin/com/xquare/featuremain/ui/Dummy.kt
@@ -1,0 +1,1 @@
+package com.xquare.featuremainapplication.ui
